requirements.txt
Flask>=2.2.5
Flask-Cors>=3.0.10
Flask-SQLAlchemy>=3.0.3
SQLAlchemy>=2.0.19
psycopg2-binary>=2.9.6
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
alembic>=1.11.1
pytest>=8.4.3
pytest-flask>=1.2.0
gunicorn>=20.1.0
